Refactor: 게시글 상세 페이지 컴포넌트 및 스타일 분리
