+++ src/pages/articleDetail/index.module.scss
@@ -87,36 +87,6 @@
     font-size: 1.1rem;
   }
 
-  .info {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    gap: 0.5rem;
-    color: $black;
-    font-weight: 400;
-    font-size: 14px;
-    margin-bottom: 2rem;
-    padding: 1rem 0.5rem;
-    border-block: 1px solid #f2f2f2;
-
-    div {
-      display: flex;
-      align-items: center;
-      gap: 0.5rem;
-    }
-
-    span {
-      display: flex;
-      align-items: center;
-      gap: 0.25rem;
-    }
-
-    .date {
-      opacity: 0.5;
-      font-size: 12px;
-    }
-  }
-
   .btns {
     display: flex;
     align-items: center;
